# openapi.yaml
openapi: 3.0.3
info:
  title: MAMV Verification API
  version: 1.0.0
  description: |
    Stable, developer-friendly API for submitting and tracking LLM verification tasks.
servers:
  - url: http://localhost:3000
paths:
  /v1/verify:
    post:
      summary: Submit a verification task
      operationId: verifyTask
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            type: string
          description: Reuse the same task when repeating a request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/tasks/{task_id}:
    get:
      summary: Fetch task status
      operationId: getTask
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/mamv/tasks/{task_id}/verify:
    get:
      summary: Verify record exists on-chain
      operationId: verifyRecordOnChain
      description: |
        Checks that a Finalized event exists on-chain for the given task.
        Returns block number and transaction hash if verified.
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: On-chain verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnChainVerifyResult'
  /api/mamv/tasks/{task_id}/receipt:
    get:
      summary: Get verification receipt for a task
      operationId: getReceipt
      description: |
        Returns the canonical verification receipt for a finalized task.
        Receipts contain all data needed for independent verification.
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Verification receipt
          content:
            application/json:
              schema:
                type: object
                properties:
                  receipt:
                    $ref: '#/components/schemas/VerificationReceipt'
        '404':
          description: Receipt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/mamv/tasks/{task_id}/receipt/verify:
    get:
      summary: Verify receipt against on-chain data
      operationId: verifyReceiptOnChain
      description: |
        Verifies that a receipt's hash matches on-chain records.
        Can optionally verify a specific receipt hash.
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
        - in: query
          name: receipt_hash
          schema:
            type: string
          description: Optional receipt hash to verify
      responses:
        '200':
          description: Receipt verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptVerifyResult'
components:
  schemas:
    ProgramIO:
      type: object
      additionalProperties: false
      required:
        - name
        - type
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
        required:
          type: boolean
    ProgramStep:
      type: object
      additionalProperties: false
      required:
        - type
      properties:
        id:
          type: string
        type:
          type: string
          enum: [prompt, retrieve, cross-check, score, evidence, consensus]
        description:
          type: string
        config:
          type: object
          additionalProperties: true
    ProgramDefinition:
      type: object
      additionalProperties: false
      required:
        - name
        - version
        - steps
      properties:
        name:
          type: string
        version:
          type: string
        description:
          type: string
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/ProgramIO'
        outputs:
          type: array
          items:
            $ref: '#/components/schemas/ProgramIO'
        steps:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ProgramStep'
...
    VerifyRequest:
      type: object
      additionalProperties: false
      required:
        - prompt
        - models
        - task_type
      properties:
        prompt:
          type: string
        models:
          type: array
          minItems: 1
          items:
            type: string
        task_type:
          type: string
          enum: [factual-qa, math-proof, policy-compliance, citation-check, general]
        deadline:
          type: integer
          minimum: 1
        commit_deadline_seconds:
          type: integer
          minimum: 1
        reveal_deadline_seconds:
          type: integer
          minimum: 1
        reward_pool:
          type: number
        program_id:
          type: string
        program:
          $ref: '#/components/schemas/ProgramDefinition'
        idempotency_key:
          type: string
    VerifyResponse:
      type: object
      additionalProperties: false
      required:
        - task_id
        - status
        - verdict
        - score_bps
        - evidence
        - timings
        - errors
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [queued, running, finalized, failed]
        verdict:
          oneOf:
            - type: boolean
            - type: string
              enum: [unknown]
        score_bps:
          type: integer
          minimum: 0
          maximum: 10000
        evidence:
          type: object
          additionalProperties: false
          required:
            - bundle_hash
            - bundle_uri
          properties:
            bundle_hash:
              type: string
              nullable: true
            bundle_uri:
              type: string
              nullable: true
        timings:
          type: object
          additionalProperties: false
          required:
            - queue_ms
            - llm_ms
            - bundle_ms
            - chain_ms
            - total_ms
          properties:
            queue_ms:
              type: number
              nullable: true
            llm_ms:
              type: number
              nullable: true
            bundle_ms:
              type: number
              nullable: true
            chain_ms:
              type: number
              nullable: true
            total_ms:
              type: number
              nullable: true
        errors:
          type: array
          items:
            $ref: '#/components/schemas/VerifyError'
        program_id:
          type: string
        program:
          $ref: '#/components/schemas/ProgramDefinition'
    VerifyError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum: [INVALID_INPUT, PROVIDER_ERROR, CHAIN_REVERT, TIMEOUT, NOT_FOUND, INTERNAL_ERROR]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/VerifyError'
...
    OnChainVerifyResult:
      type: object
      properties:
        task_id:
          type: string
        verified:
          type: boolean
          description: Whether Finalized event exists on-chain
        block_number:
          type: integer
          nullable: true
        tx_hash:
          type: string
          nullable: true
    VerificationReceipt:
      type: object
      description: |
        Canonical verification receipt - proof of task completion.
        Contains all data needed for independent verification.
      required:
        - receipt_version
        - task_id
        - generated_at
        - input_hash
        - output_hash
        - score_bps
        - verdict
        - worthy
        - evidence
        - provenance
      properties:
        receipt_version:
          type: string
          enum: ['1.0']
        task_id:
          type: string
        generated_at:
          type: integer
          description: Unix timestamp
        input_hash:
          type: string
          description: keccak256 of canonical input
        output_hash:
          type: string
          description: keccak256 of canonical output
        score_bps:
          type: integer
          minimum: 0
          maximum: 10000
        verdict:
          type: boolean
        worthy:
          type: boolean
        program:
          type: object
          nullable: true
          properties:
            program_id:
              type: string
            fingerprint:
              type: string
            name:
              type: string
            version:
              type: string
        evidence:
          type: object
          properties:
            bundle_hash:
              type: string
            bundle_uri:
              type: string
            bundle_version:
              type: string
              enum: ['0.1', '0.2']
        metering:
          type: object
          nullable: true
          properties:
            llm_calls:
              type: integer
            total_tokens:
              type: integer
            execution_ms:
              type: integer
            retrieval_calls:
              type: integer
              nullable: true
            bundle_size_bytes:
              type: integer
              nullable: true
        provenance:
          type: object
          properties:
            verifier_node:
              type: string
              nullable: true
            software_version:
              type: string
              nullable: true
            llm_provider:
              type: string
            llm_model:
              type: string
        chain_context:
          type: object
          nullable: true
          properties:
            chain_id:
              type: integer
            contract_address:
              type: string
            finalized_at:
              type: integer
            block_number:
              type: integer
            tx_hash:
              type: string
        signature:
          type: object
          nullable: true
          properties:
            signer:
              type: string
            signature:
              type: string
            signed_at:
              type: integer
        model_commitments:
          type: array
          nullable: true
          description: Model commitment hashes for accountability
          items:
            type: object
            properties:
              provider:
                type: string
              model:
                type: string
              model_commitment_hash:
                type: string
                description: keccak256 of provider+model+version+config
              inference_config_hash:
                type: string
                description: keccak256 of inference configuration
        reasoning_trace:
          type: object
          nullable: true
          description: Reasoning trace commitments (hashes only, not raw CoT)
          properties:
            trace_hash:
              type: string
              description: keccak256 of all reasoning step hashes
            trace_uri:
              type: string
              nullable: true
              description: URI to off-chain full trace (may be encrypted)
            step_count:
              type: integer
              description: Number of reasoning steps
        zk_proof:
          type: object
          nullable: true
          description: ZK proof for trustless verification (when available)
          properties:
            proof:
              type: string
              description: ZK proof bytes (hex encoded)
            public_inputs:
              type: object
              properties:
                input_hash:
                  type: string
                output_hash:
                  type: string
                model_commitment_hash:
                  type: string
                score_bps:
                  type: integer
                bundle_hash:
                  type: string
            proof_system:
              type: string
              description: Proof system used (risc_zero, sp1, noir, etc.)
    ReceiptVerifyResult:
      type: object
      properties:
        verified:
          type: boolean
        receipt_hash:
          type: string
        chain_data:
          type: object
          nullable: true
          properties:
            block_number:
              type: integer
            tx_hash:
              type: string
            finalized_at:
              type: integer
        errors:
          type: array
          items:
            type: string